User authentication

+++ routes/transaction.js
@@ -1,9 +1,9 @@
 var express = require("express");
+var secured = require("./secured.js");
 var router = express.Router();
-var pool = require("./db");
+var pool = require("../db");
 
-router.get("/get/transaction", (req, res, next) => {
-
+router.get("/get/transaction", secured(), (req, res, next) => {
   pool.query(`SELECT * FROM transaction ORDER BY merchant ASC`, (err, data) => {
     if (err) {
       next(err);
@@ -13,10 +13,9 @@ router.get("/get/transaction", (req, res, next) => {
   });
 });
 
-router.get("/get/transaction/:id", (req, res, next) => {
-  const id = parseInt(req.params.id)
-  pool.query(
-    `SELECT * FROM transaction  WHERE uid = $1`, [id], (err, data) => {
+router.get("/get/transaction/:id", secured(), (req, res, next) => {
+  const id = parseInt(req.params.id);
+  pool.query(`SELECT * FROM transaction  WHERE uid = $1`, [id], (err, data) => {
     if (err) {
       next(err);
     } else {
@@ -25,8 +24,7 @@ router.get("/get/transaction/:id", (req, res, next) => {
   });
 });
 
-
-router.post("/post/transaction", (req, res, next) => {
+router.post("/post/transaction", secured(), (req, res, next) => {
   const { merchant, payment_type, amount } = req.query;
   pool.query(
     `INSERT INTO transaction (merchant, payment_type, amount) VALUES ($1, $2, $3)`,
@@ -40,18 +38,17 @@ router.post("/post/transaction", (req, res, next) => {
       Payment: ${req.query.payment_type}
       Amount: ${req.query.amount}`);
         console.log("1 record inserted");
-        
       }
     }
   );
 });
 
-router.put("/put/transaction/:id", (req, res, next) => {
-  const id = parseInt(req.params.id)
+router.put("/put/transaction/:id", secured(), (req, res, next) => {
+  const id = parseInt(req.params.id);
   const { merchant, payment_type, amount } = req.query;
 
   pool.query(
-    'UPDATE transaction SET merchant = $1, payment_type = $2, amount = $3 ',
+    "UPDATE transaction SET merchant = $1, payment_type = $2, amount = $3 ",
     [merchant, payment_type, amount],
     (err, results) => {
       if (err) {
@@ -61,35 +58,33 @@ router.put("/put/transaction/:id", (req, res, next) => {
       Merchant: ${req.query.merchant}
       Payment: ${req.query.payment_type}
       Amount: ${req.query.amount}`);
-      console.log("1 record inserted");
-        
+        console.log("1 record inserted");
       }
     }
   );
 });
 
-router.delete("/delete/transaction/:id", (req, res, next) => {
-  const id = parseInt(req.params.id)
-  pool.query(
-    `DELETE FROM transaction WHERE uid = $1`, [id], (err, data) =>{
-      if (err) {
-        next(err);
-      } else {
-        res.status(200).send(`Transaction with the following UID was removed: ${id}`)
-      }
-    });
+router.delete("/delete/transaction/:id", secured(), (req, res, next) => {
+  const id = parseInt(req.params.id);
+  pool.query(`DELETE FROM transaction WHERE uid = $1`, [id], (err, data) => {
+    if (err) {
+      next(err);
+    } else {
+      res
+        .status(200)
+        .send(`Transaction with the following UID was removed: ${id}`);
+    }
+  });
 });
 
-router.delete("/delete/transaction/", (req, res, next) => {
-
-  pool.query(
-    `DELETE FROM transaction`, (err, data) =>{
-      if (err) {
-        next(err);
-      } else {
-        res.status(200).send(`All transactions removed`)
-      }
-    });
+router.delete("/delete/transaction/", secured(), (req, res, next) => {
+  pool.query(`DELETE FROM transaction`, (err, data) => {
+    if (err) {
+      next(err);
+    } else {
+      res.status(200).send(`All transactions removed`);
+    }
+  });
 });
 
 router.get("*", (req, res) =>
